+++ scrapyd/launcher.py
@@ -88,21 +88,34 @@ class Launcher(Service):
 # https://docs.twisted.org/en/stable/api/twisted.internet.protocol.ProcessProtocol.html
 class ScrapyProcessProtocol(protocol.ProcessProtocol):
     def __init__(self, project, spider, job, env, args):
-        self.pid = None
         self.project = project
         self.spider = spider
         self.job = job
+        self.pid = None
         self.start_time = datetime.datetime.now()
         self.end_time = None
-        self.env = env
         self.args = args
+        self.env = env
         self.deferred = defer.Deferred()
+
+    # For equality assertions in tests.
+    def __eq__(self, other):
+        return (
+            self.project == other.project
+            and self.spider == other.spider
+            and self.job == other.job
+            and self.pid == other.pid
+            and self.start_time == other.start_time
+            and self.end_time == other.end_time
+            and self.args == other.args
+            and self.env == other.env
+        )
 
     # For error messsages in tests.
     def __repr__(self):
         return (
-            f"ScrapyProcessProtocol(pid={self.pid} project={self.project} spider={self.spider} job={self.job} "
-            f"start_time={self.start_time} end_time={self.end_time} env={self.env} args={self.args})"
+            f"ScrapyProcessProtocol(project={self.project} spider={self.spider} job={self.job} pid={self.pid} "
+            f"start_time={self.start_time} end_time={self.end_time} args={self.args} env={self.env})"
         )
 
     def outReceived(self, data):